Vulnerable_Repos/02_pymetrics/requirements.txt
flask==3.0.0
psutil==5.9.6
pyyaml==6.0.1
